+++ NEWS.md
@@ -8,11 +8,21 @@
 
 ## Breaking changes
 
+* Bumped the R dependency to R (>= 3.6.0) to support the use of `grDevices::hcl.colors()` in `plotPartial()`.
+
+* Function `grid.arrange()` and the forward pipe operator `%>%` are no longer automatically imported from packages [gridExtra](https://cran.r-project.org/package=gridExtra) and [magrittr](https://cran.r-project.org/package=magrittr), respectively; users are encouraged to load them manually if needed.
+
+* Removed the `palette` and `alpha` arguments from `plotPartial()` and `autoplot()`; the latter just got absorbed into the `...` argument. By default, `plotPartial()`'s `col.regions` argument now corresponds to `grDevices::hcl.colors(100)`, which defaults to the same viridis color palette as before, just without the dependency.
+
 * `topPredictors()` is now deprecated and will be removed in the next update. Users are advised to use the [vip](https://github.com/koalaverse/vip) package instead.
 
 ## Minor changes
 
-* Switched dependency on [viridis](https://cran.r-project.org/package=viridis) to [viridisLite](https://cran.r-project.org/package=viridisLite); consequently, to keep the 'viridis' color palettes in `autoplot()`, this required bumping the [ggplot2](https://cran.r-project.org/package=ggplot2) dependency to version 3.0.0, as well as some other code tweaks under the hood [(#106)](https://github.com/bgreenwell/pdp/issues/106). 
+* Removed dependency on the retired [plyr](https://cran.r-project.org/package=plyr) package by relying directly on the [foreach](https://cran.r-project.org/package=foreach) package. Consequently, ICE curves (`ice = TRUE`) are now slightly faster to compute (since the code refactoring avoids having to post-process ICE data from wide to long format) and the corresponding progress bar (`progress = "text"`) is more honest.
+
+  * As a further consequence, the `partial()` function only supports a simple text-based progress bar (`progress = "text"`), but more options will possibly be added later.
+
+* Removed dependency on [viridis](https://cran.r-project.org/package=viridis); consequently, to keep the 'viridis' color palettes in `autoplot()`, this required bumping the [ggplot2](https://cran.r-project.org/package=ggplot2) dependency to version 3.0.0, as well as some other code tweaks under the hood [(#106)](https://github.com/bgreenwell/pdp/issues/106). 
 
 * Removed dependency on [mgcv](https://cran.r-project.org/package=mgcv) by switching to an internal C implementation of [mgcv](https://cran.r-project.org/package=mgcv)'s `in.out()` function [(#107)](https://github.com/bgreenwell/pdp/issues/107). (This is used behind the scenes whenever `partial()` is called with `chull = TRUE`.)